trade_worker/requirements.txt
psycopg[binary]==3.2.13
python-dotenv==1.1.1
